Refactor src/utils.TestJsonToResponse to test/testutils.TestJsonToResponse
