+++ test/utils_test.py
@@ -2,9 +2,8 @@
 Tests for the functions in utils module.
 """
 import pytest
-import requests
 
-from utils import extract_group_id, json_to_Response
+from utils import extract_group_id
 
 
 class TestExtractGroupId:
@@ -36,39 +35,3 @@ class TestExtractGroupId:
         with pytest.raises(ValueError):
             extract_group_id(url)
 
-
-# class TestExtractCalendarId:
-#     """Tests for the extract_calendar_id function."""
-
-#     def test_extract_calendar_id(self):
-#         """Test that the function returns the correct calendar ID."""
-#         url = "https://calendar.google.com/calendar/embed?src=univincity.fi_12345%40group.calendar.google.com&ctz=Europe%2FHelsinki"
-#         expected = "univincity.fi_12345@group.calendar.google.com"
-#         actual = extract_calendar_id(url)
-#         assert actual == expected
-
-#     def test_extract_calendar_id_empty_url(self):
-#         """Test that the function returns an empty string when the URL is empty."""
-#         url = ""
-#         expected = ""
-#         actual = extract_calendar_id(url)
-#         assert actual == expected
-
-#     def test_extract_calendar_id_no_src_param(self):
-#         """Test that the function returns an empty string when the URL does not contain the 'src' parameter."""
-#         url = "https://calendar.google.com/calendar/embed?ctz=Europe%2FHelsinki"
-#         expected = ""
-#         actual = extract_calendar_id(url)
-#         assert actual == expected
-
-
-class TestJsonToResponse:
-    """Tests for the json_to_Response function."""
-
-    def test_json_to_Response(self):
-        """Test that the function returns a Response object with the correct status code."""
-        json_file = "test/res/events-sample.json"
-        status_code = 200
-        response = json_to_Response(json_file, status_code)
-        assert isinstance(response, requests.Response)
-        assert response.status_code == status_code
